% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BothAssets}
\alias{BothAssets}
\title{Small and Mid cap assets}
\format{List with Small and Mid Cap. They have been choosen given \url{https://www.euroinvestor.dk/markeder/aktier/europa/danmark}. For each asset:
\describe{
  \item{\code{Elements in list}}{OHLCAV from yahoo Finance}
}}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
BothAssets
}
\description{
List of Small  and Mid cap assets prices from Yahoo Finance
}
\keyword{datasets}
