% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MidAssets}
\alias{MidAssets}
\title{MidAssets}
\format{List with Mid  Cap. They have been choosen given \url{https://www.euroinvestor.dk/markeder/aktier/europa/danmark}. For each asset:
\describe{
  \item{\code{Elements in list}}{OHLCAV from yahoo Finance}
}}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
MidAssets
}
\description{
List of Mid cap assets prices from Yahoo Finance
}
\keyword{datasets}
