% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SmallAssets}
\alias{SmallAssets}
\title{SmallAssets}
\format{List with Small Cap. They have been choosen given \url{https://www.euroinvestor.dk/markeder/aktier/europa/danmark}. For each asset:
\describe{
  \item{\code{Elements in list}}{OHLCAV from yahoo Finance}
}}
\source{
\url{https://finance.yahoo.com/}
}
\usage{
SmallAssets
}
\description{
List of Small cap assets prices from Yahoo Finance
}
\keyword{datasets}
